完成重构, demo1.0
