+++ Draw.cs
@@ -18,6 +18,102 @@ namespace homework
     public class Draw
     {
         /// <summary>
+        /// 建图节点
+        /// </summary>
+        struct graph_node
+        {
+            public int id;
+            public double x;
+            public double y;
+            public graph_node(int id, double x, double y)
+            {
+                this.id = id;
+                this.x = x;
+                this.y = y;
+            }
+        }
+        
+        /// <summary>
+        /// 初始界面绕圈绘图函数，
+        /// </summary>
+        /// <param name="playground">画布的引用</param>
+        /// <param name="graph">图的邻接表</param>
+        /// <param name="directed">是否为有向图</param>
+        /// <param name="count">节点数</param>
+        public static void Generate_Graph(ref Canvas playground, Graph graph, bool directed, int count)
+        {
+            playground.Children.Clear();
+            double R = 150; double x0 = 283; double y0 = 194; double r = 25;
+            double du = 2 * 3.1415926535 / count;
+            List<graph_node> graph_data = new List<graph_node>();
+            graph_data.Add(new graph_node());
+
+            //画圆
+            for (int i = 1; i <= count; i++)
+            {
+                Ellipse e = new Ellipse();
+                TextBlock l = new TextBlock();
+                l.Text = Convert.ToString(i);
+                l.FontSize = 20;
+                double y = y0 - Math.Cos(du * i) * R;
+                double x = x0 + Math.Sin(du * i) * R;
+                //此处x,y并非圆心，画图可知为⚪的交点
+                //MessageBox.Show("创建位置：" + Convert.ToString(x - r) + " " + Convert.ToString(y - r));
+                Canvas.SetLeft(e, x - r / 2); Canvas.SetTop(e, y - r / 2);
+                Canvas.SetLeft(l, x); Canvas.SetTop(l, y);
+                Canvas.SetZIndex(e, 1);
+                Canvas.SetZIndex(l, 1);
+                graph_data.Add(new graph_node(i, x + r / 2, y + r / 2));
+                e.Width = e.Height = 2 * r;
+                e.Fill = new SolidColorBrush(Color.FromRgb(108, 165, 178));
+                playground.Children.Add(e);
+                playground.Children.Add(l);
+            }
+
+            List<List<Edge>> nodeslist = graph.getList();
+
+            //画直线和箭头
+            for (int i = 1; i <= nodeslist.Count - 1; i++)
+            {
+                for (int j = 1; j <= nodeslist[i].Count - 1; j++)
+                {
+                    Line l = new Line(); // 该直线为图中连线
+                    l.Stroke = Brushes.Black;
+                    l.StrokeThickness = 3;
+                    //指定起点和终点的坐标
+                    l.X1 = graph_data[i].x; l.Y1 = graph_data[i].y;
+                    l.X2 = graph_data[nodeslist[i][j].t].x; l.Y2 = graph_data[nodeslist[i][j].t].y;
+                    Canvas.SetZIndex(l, 0);
+
+                    ///////////---------------以下为确定箭头终点-----------------------------------------------------//////////////
+
+                    double k = (l.Y2 - l.Y1) / (l.X2 - l.X1);
+                    //- -   右半区+下顶点正确
+                    //即  l.x1 < l.x2 时结果正确
+                    //+ +  左半区+上顶点正确
+                    //即  l.x1 > l.x2 时结果正确
+
+                    double xm = l.X2;
+                    double ym = l.Y2;
+
+                    if (l.X1 < l.X2)
+                    {
+                        xm -= r * Math.Cos(Math.Atan(k));
+                        ym -= r * Math.Sin(Math.Atan(k));
+                    }
+                    else
+                    {
+                        xm += r * Math.Cos(Math.Atan(k));
+                        ym += r * Math.Sin(Math.Atan(k));
+                    }
+
+                    playground.Children.Add(l);
+                    if (directed == true) Draw.DrawArrow(ref playground, xm, ym, Draw.PI / 6, 10, l);
+                }
+            }
+        }
+
+        /// <summary>
         /// 圆周率
         /// </summary>
         public static double PI = 3.1415926535;
